### 2025-08-04 #8735 @arfon
@pcauchy1 – please consult the documentation and reformat your paper to _exactly_ match https://joss.readthedocs.io/en/latest/example_paper.html
