+++ paper.bib
@@ -1,0 +1,95 @@
+@article{wilkinson:2016,
+  title = {The FAIR Guiding Principles for scientific data management and stewardship},
+  author = {Wilkinson, Mark D. and Dumontier, Michel and Aalbersberg, IJsbrand Jan and others},
+  journal = {Scientific Data},
+  volume = {3},
+  year = {2016},
+  doi = {10.1038/sdata.2016.18}
+}
+
+@article{savitzky:2021,
+  title = {py4DSTEM: a software package for four-dimensional scanning transmission electron microscopy data analysis},
+  author = {Savitzky, Benjamin H. and others},
+  journal = {Microscopy and Microanalysis},
+  volume = {27},
+  pages = {712},
+  year = {2021},
+  doi = {10.1017/S1431927621003251}
+}
+
+@software{pena:2025,
+  title = {Hyperspy/Hyperspy: V2.3.0},
+  author = {Peña, Francisco de la and others},
+  year = {2025},
+  publisher = {Zenodo},
+  doi = {10.5281/zenodo.14956374}
+}
+
+@software{wojdyr:2025,
+  title = {Wojdyr/Xylib},
+  author = {Wojdyr, Marcin},
+  year = {2025},
+  howpublished = {\url{https://github.com/wojdyr/xylib}}
+}
+
+@software{htmdec:2025,
+  title = {Htmdec/Htmdec_formats},
+  author = {{HTMDEC}},
+  year = {2025},
+  howpublished = {\url{https://github.com/htmdec/htmdec_formats}}
+}
+
+@misc{mgi:2021,
+  title = {Materials Genome Initiative Strategic Plan},
+  author = {{U.S. White House Office of Science and Technology Policy}},
+  year = {2021},
+  howpublished = {\url{https://www.mgi.gov/sites/default/files/documents/MGI-2021-Strategic-Plan.pdf}},
+  note = {Accessed: 2025-06-13}
+}
+
+@article{brinson:2024,
+  title = {Community action on FAIR data will fuel a revolution in materials research},
+  author = {Brinson, L. Catherine and Bartolo, Laura M. and Blaiszik, Ben and others},
+  journal = {MRS Bulletin},
+  volume = {49},
+  pages = {12--16},
+  year = {2024},
+  doi = {10.1557/s43577-023-00498-4}
+}
+
+@article{odegard:2023,
+  title = {A successful strategy for MGI-inspired research},
+  author = {Odegard, Gregory M. and Liang, Zhen and Siochi, Emilie J. and others},
+  journal = {MRS Bulletin},
+  volume = {48},
+  pages = {434--438},
+  year = {2023},
+  doi = {10.1557/s43577-023-00525-4}
+}
+
+@article{depablo:2019,
+  title = {New frontiers for the materials genome initiative},
+  author = {de Pablo, Juan J. and Jackson, Nicholas E. and Webb, Michael A. and others},
+  journal = {npj Computational Materials},
+  volume = {5},
+  pages = {41},
+  year = {2019},
+  doi = {10.1038/s41524-019-0173-4}
+}
+
+@book{nasem:2024,
+  title = {Foundational Research Gaps and Future Directions for Digital Twins},
+  author = {{National Academies of Sciences, Engineering, and Medicine}},
+  year = {2024},
+  publisher = {The National Academies Press},
+  address = {Washington, DC},
+  doi = {10.17226/26894}
+}
+
+@misc{paradim:2025,
+  title = {PARADIM Data Portal},
+  author = {{Platform for the Accelerated Realization, Analysis, and Discovery of Interface Materials (PARADIM)}},
+  year = {2025},
+  howpublished = {\url{https://data.paradim.org/index.php}},
+  note = {Accessed: 2025-06-13}
+}
