+++ paper.md
@@ -1,67 +1,58 @@
-
-# Project Chameleon: A Python Package for Interoperability in Experimental Materials Science Research Data
-
-## tags:
+---
+title: 'Project Chameleon: A Python Package for Interoperability in Experimental Materials Science Research Data'
+tags:
   - Python
-  -  FAIR Data
-  -  materials science
+  - FAIR Data
+  - materials science
   - crystal growth
   - reflection high-energy electron diffraction
   - angle-resolved photoemission spectroscopy
   - molecular beam epitaxy
   - scanning transmission electron microscopy
   - x-ray diffraction
-
-# Authors:
-  
-- Peter Cauchy<sup>1, 2, 3, 4</sup> (https://orcid.org/0009-0007-1155-0509) 
-- David Elbert<sup>1, 2</sup> (https://orcid.org/0000-0002-2292-180X)
-- Tyrel McQueen<sup>1, 2, 3, 4</sup> (https://orcid.org/0000-0002-8493-4630)
-
-# Affiliations:
-1. PARADIM, William H. Miller III Department of Physics and Astronomy, Johns Hopkins University, Baltimore 21218, MD (https://ror.org/04pw1zg89)
-2. Hopkins Extreme Materials Institute, Johns Hopkins University, Baltimore, MD 21218 (https://ror.org/02ed2th17)
-3. Department of Chemistry, Johns Hopkins University, Baltimore 21218, MD, 
-4. Department of Materials Science and Engineering, Johns Hopkins University, Baltimore 21218, MD, 
-
-
+authors:
+  - name: Peter Cauchy
+    orcid: 0009-0007-1155-0509
+    affiliation: "1, 2, 3, 4"
+  - name: David Elbert
+    orcid: 0000-0002-2292-180X
+    affiliation: "1, 2"
+  - name: Tyrel McQueen
+    orcid: 0000-0002-8493-4630
+    affiliation: "1, 2, 3, 4"
+affiliations:
+ - name: PARADIM, William H. Miller III Department of Physics and Astronomy, Johns Hopkins University, Baltimore 21218, MD
+   index: 1
+   ror: 04pw1zg89
+ - name: Hopkins Extreme Materials Institute, Johns Hopkins University, Baltimore, MD 21218
+   index: 2
+   ror: 02ed2th17
+ - name: Department of Chemistry, Johns Hopkins University, Baltimore 21218, MD
+   index: 3
+ - name: Department of Materials Science and Engineering, Johns Hopkins University, Baltimore 21218, MD
+   index: 4
+date: 13 August 2025
+bibliography: paper.bib
+---
 
 # Summary
 
-Project Chameleon is a versatile Python package for extensible data transformation and interoperability. Chameleon streamlines analysis and combination of scientific data from different sources and provides a foundational pillar for implementation of the FAIR (Findable, Accessible, Interoperable, and Reusable) data principles[1]. In the context of FAIR, interoperability refers to the ability of diverse datasets, tools, and systems to seamlessly interact, integrate, and exchange data. Interoperability ensures that scientific data can be readily combined or used alongside related datasets, software, and analytical tools across different domains and disciplines.  A central barrier to interoperability arises from heterogeneous data formats and differing metadata conventions encoded in the varied file formats of different laboratories.  Project Chameleon transformations eliminate this barrier to provide effective interoperability that facilitates comprehensive analyses and machine learning that accelerate scientific discovery and enhance reproducibility of scientific findings.
[…]
