### 2025-10-02 paradimdata/project_chameleon#32 @mxwalbert
[JOSS Review] Installation instructions unclear
Dear authors,

I identified some issues with your installation instructions:

- The steps provided in [Installation and Requirements](https://project-chameleon.readthedocs.io/en/latest/introduction/installing_chameleon.html#installation-and-requirements) are inconsistent. There is a mix up of multiple package managers (`pipx`, `pip`, `poetry`, `conda`) which could lead to mismatched installs.
- The installation instructions for the required compiler packages `boost` and `swig` are only provided for usage with `brew`. Manual instructions are provided further down on the page but seem quite cumbersome.
- Notes on using the [Docker Image](https://project-chameleon.readthedocs.io/en/latest/introduction/installing_chameleon.html#docker-image) are provided but clear instructions are left out. However, a step-by-step guide is available in [Chameleon Docker Image](https://project-chameleon.readthedocs.io/en/latest/implementation/docker.html), so they could just be linked there.
- The Girder plugin is not available under the provided link: [Chameleon plugin repository](https://github.com/paradimdata/chameleon_plugin)

I think a solution to resolve the first two points is to base the whole installation on conda, which could also be used for pulling `boost` and `swig`.  Then, provided a requirements file, the installation could be simplified to a single line of code.

This issue is related to openjournals/joss-reviews#8963
