### 2025-10-02 paradimdata/project_chameleon#33 @mxwalbert
[JOSS Review] Incorrect licensing
Dear authors,

Your package is distributed under the MIT license. However, the repository contains at least one file ([xyconv.py](https://github.com/paradimdata/project_chameleon/blob/main/project_chameleon/xyconv.py)) which is copied from a LGPL-2.1-licensed package ([`xylib`](https://github.com/wojdyr/xylib)). As far as I can tell, this is not permitted because MIT is more permissive and would strip away the LGPL’s copyleft requirements. You may remove this file (since I do not see where it is used in your package anyway) or adapt your license accordingly. Also, please make sure that there are no similar issues in your repository.

This issue is related to openjournals/joss-reviews#8963
