### 2025-10-02 paradimdata/project_chameleon#34 @mxwalbert
[JOSS Review] Community guidelines missing
Dear authors,

The community guidelines are missing in the repository and the documentation.

There should be clear guidelines for third-parties wishing to:
- Contribute to the software
- Report issues or problems with the software
- Seek support

This issue is related to openjournals/joss-reviews#8963
