### 2025-10-02 paradimdata/project_chameleon#35 @mxwalbert
[JOSS Review] Tests fail
Dear authors,

Thank you for adding tests to the repository to quickly check the functionality of your software.

Unfortunately, a large number of tests fail after following the manual installation instructions as well as when using the Docker image:
- Some of the tests fail to load due to missing dependencies.
- Most of the running tests seem to fail due to issues with locating the correct path.

This issue is related to openjournals/joss-reviews#8963
