### 2025-10-02 paradimdata/project_chameleon#36 @mxwalbert
[JOSS Review] Documentation incomplete
Dear authors,

Thank you for providing a reference for the implemented functions in the documentation.

However, the `FastAPI`-based app, which is advertised as a feature in the JOSS paper, is not sufficiently documented.

Additionally, the term "REST API" might be confusing since the endpoints are remote procedure calls rather than resources in the classical sense. Correct terms in this case could be "web API" or "HTTP API".

This issue is related to openjournals/joss-reviews#8963
