+++ project_chameleon/brukerrawbackground.py
@@ -1,6 +1,5 @@
 from .brukerrawconverter import brukerrawconverter
 import pandas as pd
-import numpy as np
 import matplotlib.pyplot as plt
 from pathlib import Path
 import csv
