+++ project_chameleon/mbeparser.py
@@ -3,7 +3,6 @@ import pathlib
 import shutil
 import matplotlib.pyplot as plt
 import numpy as np
-import sys
 import argparse
 
 def find_shutter_values(shutter_input):
