+++ project_chameleon/rheed_video_converter.py
@@ -1,9 +1,5 @@
 import os
-import glob
-import numpy as np
-from PIL import Image
 import argparse
-import shutil
 from sys import argv, exit
 from time import time, sleep
 from subprocess import Popen, PIPE, DEVNULL
