### 2026-02-09 cba91d4
Update index.rst
