### 2026-02-09 8049a92
Update installing_chameleon.rst
