+++ environment.yml
@@ -1,0 +1,29 @@
+name: project-chameleon
+channels:
+  - conda-forge
+dependencies:
+  # Core
+  - python>=3.12.2
+  - pip
+
+  # System / compiled dependencies
+  - boost
+  - swig
+
+  # Python dependencies (from pyproject.toml)
+  - sphinx>=8.1.3
+  - matplotlib>=3.9.0
+  - numpy>=1.26.4
+  - hyperspy>=2.1.0
+  - py4dstem>=0.14.15
+  - pandas>=2.2.2
+  - sphinx-rtd-theme>=1.0.0
+  - imagehash>=4.3.1
+
+  # Dev / test dependencies
+  - pytest>=8.2.2
+
+  # Install the project itself
+  - pip:
+      - xylib-py>=1.6.1
+      - .
