+++ environment.yml
@@ -19,6 +19,7 @@ dependencies:
   - pandas>=2.2.2
   - sphinx-rtd-theme>=1.0.0
   - imagehash>=4.3.1
+  - openpyxl>=3.1
 
   # Dev / test dependencies
   - pytest>=8.2.2
