### 2026-02-10 5f3ef2f
Update installing_chameleon.rst
