### 2026-02-10 6de97f8
Update plugin.rst
