### 2026-02-10 #8963 @pcauchy1
Hey @rkurchin all issues have been addressed. Sorry for the delay. Here are the changes made and the issues the correspond to:

1. Functionality Installation
- installation method changed to automatically install boost and swig.
- Further documentation added to RTD so there is less back and forth between github repo and RTD.

2. Documentation Incomplete
- Added documentation specifically for web API
- Terminology changed

3. Tests Fail
- Resolved issues with paths in test

4.Community Guidlines Missing
- Added community guidelines to RTD home page

5. Incorrect licensing
- Removed file causing the issue

6. Installation instructions Unclear
- As mentioned above, installation simplified
- Appropriate links added

7. State of the field
- Related work section added

Let me know what the next steps are from here or if there is anything else you need from me.
