### 2026-03-06 #8963 @mxwalbert
Dear @rkurchin,
Thank you for the reminder!

Dear @pcauchy1,
Unfortunately, the recommended installation via Miniconda3 still does not work for me when I follow the steps in the documentation. There is an error with building xylib-py. I am using Windows 11 on a fresh Miniconda3 installation. Therefore, I am not able to test the functionality of the package.
In general, I would ask you to put your comments from above inside the related [issues on your repository](https://github.com/paradimdata/project_chameleon/issues) and link the commits where you addressed them. This will help to keep track of what needs to be done, since resolved issues can be closed.

Still, these issues are only of formal nature. I also raised **concern on the scope eligibility** in [my previous comment](https://github.com/openjournals/joss-reviews/issues/8963#issuecomment-3363373086):
> However, I also need to raise a concern about the general scope of the package. The functionality seems to be restricted to a specific collection of measurement equipment and the applicability in other labs is questionable. In the documentation you and your co-authors state yourself multiple times that the provided functions are "specifically designed" for a certain instrument. Moreover, it seems like a big part of the logic is done by third-party dependencies and the implemented functions mainly perform data formatting.

@pcauchy1 please comment on the above statement.
