### 2026-08-17 #8963 @Grant-Xue
@rkurchin Sorry for the late reply. I was busy with a funding application last month.
I have now opened four issues covering installation, automated tests, the software paper, and functionality documentation:
- paradimdata/project_chameleon#38
- paradimdata/project_chameleon#39
- paradimdata/project_chameleon#40
- paradimdata/project_chameleon#41
 I am currently continuing with the functionality testing and will update the review accordingly.
